backend (favicon, menu sidebar, dashboard, etc)

+++ database/migrations/2023_01_06_075122_create_purchases_table.php
@@ -18,7 +18,7 @@ class CreatePurchasesTable extends Migration
             $table->integer('supplier_id')->index();
             $table->dateTime('purchase_date');
             $table->string('recipient');            
-            $table->integer('amount');            
+            $table->double('amount');            
             $table->string('code')->nullable();            
             $table->text('description')->nullable();            
             $table->timestamps();
